+++ mod-infra/src/main/scala/at/forsyte/apalache/infra/DefaultExceptionAdapter.scala
@@ -7,5 +7,9 @@ import javax.inject.Singleton
   */
 @Singleton
 class DefaultExceptionAdapter extends ExceptionAdapter {
-  override def toMessage: PartialFunction[Exception, String] = PartialFunction.empty
+  /**
+    * Given an exception, the adapter produces an error message along with its severity.
+    * @return
+    */
+  override def toMessage: PartialFunction[Exception, ErrorMessage] = PartialFunction.empty
 }